restructured types, funcs
